#20: move GenerateXYZ test to its own file
